+++ third_party_libs/FFTConvolver/wrapper.cpp
@@ -14,19 +14,21 @@ int NumFrames(StereoConvolver& convolver) { return convolver.num_frames; }
 
 void DestroyStereoConvolver(StereoConvolver* convolver) { delete convolver; }
 
-void Init(StereoConvolver& convolver, float const* samples, int num_frames, int num_channels) {
+void Init(StereoConvolver& convolver, float const* samples, float gain_db, int num_frames, int num_channels) {
     convolver.num_frames = num_frames;
 
     auto channel_samples = new float[(unsigned)num_frames];
 
     assert(num_channels == 1 || num_channels == 2);
 
+    float gain = powf(10.0f, gain_db / 20.0f);
+
     // IMPROVE: probably more efficent ways to handle mono audio
 
     for (int chan = 0; chan < 2; ++chan) {
         if (!(chan == 2 && num_channels == 1))
             for (int frame = 0; frame < num_frames; ++frame)
-                channel_samples[frame] = samples[frame * num_channels + chan];
+                channel_samples[frame] = samples[frame * num_channels + chan] * gain;
 
         // Just trial and error with these values; these seem to be most efficient
         constexpr size_t k_head_block_size = 512;